+++ src/drivel/cli/constants.py
@@ -3,10 +3,20 @@ from pathlib import Path
 from auto_name_enum import AutoNameEnum, auto
 
 
-class OutputFormat(AutoNameEnum):
+class ThemeListOutputFormat(AutoNameEnum):
     json = auto()
     spaces = auto()
     lines = auto()
 
 
-CACHE_DIR: Path = Path.home() / ".local/share/meta-fun"
+class ThemeOutputFormat(AutoNameEnum):
+    json = auto()
+    yaml = auto()
+
+
+SHARE_DIR: Path = Path.home() / ".local/share/drivel"
+CONFIG_DIR: Path = Path.home() / ".config/drivel"
+CACHE_DIR: Path = Path.home() / ".cache/drivel"
+
+SETTINGS_FILE_NAME: str = "settings.json"
+EXTRA_THEMES_DIR: str = "extra-themes"