+++ src/drivel/cli/format.py
@@ -1,5 +1,6 @@
 import json
-from typing import Any, Iterable
+from typing import Any
+from collections.abc import Iterable
 
 import pyperclip
 import snick
@@ -7,9 +8,10 @@ from loguru import logger
 from rich.console import Console
 from rich.panel import Panel
 from rich.markdown import Markdown
+import yaml
 
 
-def _to_clipboard(text) -> bool:
+def _to_clipboard(text: str) -> bool:
     try:
         pyperclip.copy(text)
         return True
@@ -19,13 +21,13 @@ def _to_clipboard(text) -> bool:
 
 
 def terminal_message(
-    message,
-    subject=None,
-    color="green",
-    footer=None,
-    indent=True,
-    markdown=False,
-    to_clipboard=False,
+    message: str,
+    subject: str | None = None,
+    color: str = "green",
+    footer: str | None = None,
+    indent: bool = True,
+    markdown: bool = False,
+    to_clipboard: bool = False,
 ):
     if to_clipboard:
         result = _to_clipboard(message)
@@ -39,36 +41,38 @@ def terminal_message(
     text = snick.dedent(message)
     if indent:
         text = snick.indent(text, prefix="  ")
-    if markdown:
-        text = Markdown(text)
     console = Console()
     console.print()
-    console.print(Panel(text, **panel_kwargs))
+    if markdown:
+        console.print(Panel(Markdown(text), **panel_kwargs))
+    else:
+        console.print(Panel(text, **panel_kwargs))
     console.print()
 
 
 def simple_message(
-    message,
-    indent=False,
-    markdown=False,
-    to_clipboard=False,
+    message: str,
+    indent: bool = False,
+    markdown: bool = False,
+    to_clipboard: bool = False,
 ):
     text = snick.dedent(message)
     if indent:
         text = snick.indent(text, prefix="  ")
-    if markdown:
-        text = Markdown(text)
     if to_clipboard:
         result = _to_clipboard(text)
         if result:
             logger.debug("output copied to clipboard")
     console = Console()
     console.print()
-    console.print(text)
+    if markdown:
+        console.print(Markdown)
+    else:
+        console.print(text)
     console.print()
 
 
-def as_spaces(stuff: Iterable, fancy: bool = False, **kwargs):
+def as_spaces(stuff: Iterable[Any], fancy: bool = False, **kwargs: Any):
     output = " ".join(stuff)
     if fancy:
         return terminal_message(output, **kwargs)
@@ -77,7 +81,7 @@ def as_spaces(stuff: Iterable, fancy: bool = False, **kwargs):
     return simple_message(output, **kwargs)
 
 
-def as_lines(stuff: Iterable, fancy: bool = False, **kwargs):
+def as_lines(stuff: Iterable[Any], fancy: bool = False, **kwargs: Any):
     output = "\n".join(stuff)
     if fancy:
         return terminal_message(output, **kwargs)
@@ -86,7 +90,7 @@ def as_lines(stuff: Iterable, fancy: bool = False, **kwargs):
     return simple_message(output, **kwargs)
 
 
-def as_json(stuff, fancy: bool = False, to_clipboard=False, **kwargs):
+def as_json(stuff: Any, fancy: bool = False, to_clipboard: bool = False, **kwargs: Any):
     indent: int | None = 2 if fancy else None
     if to_clipboard:
         result = _to_clipboard(json.dumps(stuff, indent=indent))
@@ -95,4 +99,18 @@ def as_json(stuff, fancy: bool = False, to_clipboard=False, **kwargs):
     console = Console()
     console.print()
     console.print_json(data=stuff, indent=indent, **kwargs)
+    console.print()
+
+
+def as_yaml(stuff: Any, fancy: bool = False, to_clipboard: bool = False, **kwargs: Any):
+    if fancy:
+        logger.warning("as_yaml doesn't support fancy output")
+    text = yaml.dump(stuff, sort_keys=False)
+    if to_clipboard:
+        result = _to_clipboard(text)
+        if result:
+            logger.debug("output copied to clipboard")
+    console = Console()
+    console.print()
+    console.print(text, **kwargs)
     console.print()