+++ src/drivel/cli/give.py
@@ -1,11 +1,13 @@
+from pathlib import Path
 from typing import Annotated, Any
 
 import typer
 
-from drivel.cli.cache import init_cache
+from drivel.exceptions import DrivelError
 from drivel.cli.config import Settings, attach_settings
-from drivel.cli.constants import OutputFormat
-from drivel.cli.exceptions import handle_abort
+from drivel.cli.constants import ThemeListOutputFormat
+from drivel.cli.exceptions import Abort, handle_abort
+from drivel.cli.extra_themes import attach_extra_themes
 from drivel.cli.format import as_spaces, as_lines, as_json
 from drivel.themes import Theme
 
@@ -14,9 +16,9 @@ cli = typer.Typer()
 
 
 @cli.callback(invoke_without_command=True)
+@attach_settings(validate=True)
+@attach_extra_themes
 @handle_abort
-@init_cache
-@attach_settings
 def give(
     ctx: typer.Context,
     max_count: Annotated[int | None, typer.Argument(help="The maximum number of metasyntactic names to give")] = None,
@@ -24,10 +26,10 @@ def give(
     theme_name: Annotated[
         str | None, typer.Option("--theme", help="The theme to use (If not provided, will use current default")
     ] = None,
-    kind: Annotated[str, typer.Option(help="The kind of names to give. Use 'all' to pull from all kinds")] = "default",
+    kind: Annotated[str | None, typer.Option(help="The kind of names to give. Use 'all' to pull from all kinds")] = None,
     output_format: Annotated[
-        OutputFormat, typer.Option("--format", help="The output format to use")
-    ] = OutputFormat.spaces,
+        ThemeListOutputFormat, typer.Option("--format", help="The output format to use")
+    ] = ThemeListOutputFormat.spaces,
     fancy: Annotated[bool, typer.Option(help="Enable fancy output")] = True,
     to_clipboard: Annotated[bool, typer.Option(help="Copy output to clipboard")] = True,
 ):
@@ -35,22 +37,30 @@ def give(
     Give N fun metasyntactic variable names.
     """
     settings: Settings = ctx.obj.settings
+    extra_themes_dir: Path = ctx.obj.extra_themes
     if theme_name is None:
         theme_name = settings.default_theme
-    theme = Theme.load(theme_name)
-    items = theme.give(max_count=max_count, kind=kind, do_shuffle=do_shuffle)
+
+    try:
+        theme = Theme.load(theme_name, extra_themes_dir)
+        items = theme.give(max_count=max_count, kind=kind, do_shuffle=do_shuffle)
+    except DrivelError:
+        raise Abort(
+            "There was an error fetching names!",
+            subject="Error giving names",
+        )
     kwargs: dict[str, Any] = dict(
         fancy=fancy,
         to_clipboard=to_clipboard,
     )
     match output_format:
-        case OutputFormat.spaces | OutputFormat.lines:
+        case ThemeListOutputFormat.spaces | ThemeListOutputFormat.lines:
             if fancy:
                 kwargs["subject"] = f"Fun names from {theme_name}"
             match output_format:
-                case OutputFormat.spaces:
+                case ThemeListOutputFormat.spaces:
                     as_spaces(items, **kwargs)
-                case OutputFormat.lines:
+                case ThemeListOutputFormat.lines:
                     as_lines(items, **kwargs)
-        case OutputFormat.json:
+        case ThemeListOutputFormat.json:
             as_json(items, **kwargs)