+++ src/drivel/themes.py
@@ -1,15 +1,16 @@
+from importlib.abc import Traversable
 import re
-from collections.abc import Generator
 from pathlib import Path
 from random import shuffle
+from typing import Annotated
 
 import yaml
 from auto_name_enum import AutoNameEnum, auto
 from loguru import logger
-from pydantic import BaseModel, model_validator
+from pydantic import BaseModel, Field, model_validator
 
 from drivel.constants import DEFAULT_THEME
-from drivel.exceptions import ThemeNotFound, ThemeReadError
+from drivel.exceptions import ThemeNotFound, ThemeReadError, DuplicateTheme, ThemeWriteError
 from drivel.utilities import asset_root
 
 
@@ -30,7 +31,7 @@ class ThemeMetadata(BaseModel):
 
 
 class Theme(BaseModel):
-    name: str
+    name: Annotated[str, Field(exclude=True)]
     default: str
     kinds: Kinds
     metadata: ThemeMetadata
@@ -52,14 +53,16 @@ class Theme(BaseModel):
     def give(
         self,
         max_count: int | None = None,
-        kind: str = "default",
+        kind: str | None = None,
         do_shuffle: bool = False,
     ) -> list[str]:
         logger.debug(f"Giving items from {self.name} for kind '{kind}'")
+        if kind is None:
+            kind = self.default
         if kind == "all":
             items = [i for k in self.kinds.values() for i in k]
         else:
-            items = self.kinds[self.default]
+            items = self.kinds[kind]
         if do_shuffle:
             shuffle(items)
         if max_count is None:
@@ -67,7 +70,7 @@ class Theme(BaseModel):
         return items[:max_count]
 
     @staticmethod
-    def _find(name: str, *search_paths: Path) -> Path:
+    def _find(name: str, *search_paths: Path | Traversable) -> Path | Traversable:
         for search_path in search_paths:
             for path in search_path.iterdir():
                 if path.is_file() and path.name == f"{name}.yaml":
@@ -75,32 +78,48 @@ class Theme(BaseModel):
         raise ThemeNotFound(f"Theme '{name}' not found in paths: {search_paths}")
 
     @classmethod
-    def load(cls, name: str = DEFAULT_THEME, *extra_paths: Path) -> "Theme":
-        builtin_path = Path(str(asset_root))
-
-        path = cls._find(name, builtin_path, *extra_paths)
-
-        with ThemeReadError.handle_errors(f"Couldn't read theme '{name}' from {path}"):
-            text = path.read_text()
-            data = yaml.safe_load(text)
-            theme = Theme(name=name, **data)
+    def loads(cls, text: str, name: str) -> "Theme":
+        data = yaml.safe_load(text)
+        theme = Theme(name=name, **data)
 
         return theme
 
     @classmethod
-    def names(cls, *extra_paths: Path) -> set[str]:
-        builtin_path = Path(str(asset_root))
+    def load(cls, name: str = DEFAULT_THEME, *extra_paths: Path) -> "Theme":
+        path = cls._find(name, asset_root, *extra_paths)
 
+        with ThemeReadError.handle_errors(f"Couldn't read theme '{name}' from {path}"):
+            text = path.read_text()
+            theme = cls.loads(text, name)
+
+        return theme
+
+    def dump(self, target: Path) -> None:
+        yaml_text = yaml.dump(self.model_dump())
+        theme_path = target / f"{self.name}.yaml"
+        with ThemeWriteError.handle_errors(f"Couldn't write theme '{self.name}' to {theme_path}"):
+            theme_path.write_text(yaml_text)
+
+
+    @classmethod
+    def names(cls, *extra_paths: Path) -> set[str]:
         names: set[str] = set()
-        for asset in builtin_path.iterdir():
+        for asset in asset_root.iterdir():
             if not asset.is_file():
                 continue
-            filename = Path(asset.name)
-            if filename.suffix != ".yaml":
+            (name, ext) = asset.name.rsplit('.', 1)
+            if ext != "yaml":
                 continue
-            name: str = ThemeError.enforce_defined(filename.stem, f"Invalid theme file found in {builtin_path}: {filename}")
-            set.add(names, filename.stem)
+            names.add(name)
 
-        ThemeError.require_condition(name not in names, f"Duplicate theme name found: {name}")
+        for extra_path in extra_paths:
+            for asset in extra_path.iterdir():
+                if not asset.is_file():
+                    continue
+                if asset.suffix != ".yaml":
+                    continue
+                name = asset.stem
+                DuplicateTheme.require_condition(name not in names, f"Duplicate theme name found in {extra_path}: {name}")
+                names.add(name)
 
         return names