+++ CHANGELOG.md
@@ -5,5 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.2.0 - 2025-03-21
+* Added cache management and decorators.
+* Added share management and decorators.
+* Added capability to add user themes.
+* Added ability to show themes and the theme schema.
+
+* Improved error management.
+* Improved decorators for config.
+
+* Added a lot more content to README.md
+* Added type checking
+
 ## v0.1.0 - 2025-03-21
 * Initial release